refactor: rename ng-subscriptions folder to subscriptions
https://gravitee.atlassian.net/browse/APIM-2182
