+++ gravitee-apim-console-webui/src/management/api/portal/subscriptions/components/api-key-validation/api-key-validation.harness.ts
