+++ gravitee-apim-console-webui/src/management/api/portal/subscriptions/components/dialogs/reject/api-portal-subscription-reject-dialog.component.scss
