+++ gravitee-apim-console-webui/src/management/api/portal/subscriptions/components/dialogs/transfer/api-portal-subscription-transfer-dialog.component.scss
