+++ gravitee-apim-console-webui/src/management/api/portal/subscriptions/components/dialogs/validate/api-portal-subscription-validate-dialog.component.scss
