+++ gravitee-apim-console-webui/src/management/api/portal/subscriptions/edit/api-portal-subscription-edit.component.html
